+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Updates/Fixes
 
 - Fix addresss indexing issue with ledger [#300](https://github.com/asgardex/asgardex-desktop/pull/300)
-- Fix bsc usdt not showing [#304] ()
-- Fix usdt usdt crash [#305] ()
+- Fix bsc usdt not showing [#304] (https://github.com/asgardex/asgardex-desktop/pull/305)
+- Fix usdt usdt crash [#305] (https://github.com/asgardex/asgardex-desktop/pull/306)
 
 ## Add