+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix addresss indexing issue with ledger [#300](https://github.com/asgardex/asgardex-desktop/pull/300)
 - Fix BSC.USDT not showing [#303](https://github.com/asgardex/asgardex-desktop/pull/304)
+- Fix white screen on bondtab [#301] (https://github.com/asgardex/asgardex-desktop/pull/306)
 
 # 1.21.9 (2024-06-15)